+++ galaxies/app/Http/Controllers/Api/PlanetController.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePlanetRequest;
 use App\Http\Resources\PlanetResource;
 use App\Models\Planet;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class PlanetController extends Controller
 {
@@ -17,7 +17,14 @@ class PlanetController extends Controller
      */
     public function index()
     {
-        return PlanetResource::collection(Planet::all());
+        $userId = auth('api')->user()->getAuthIdentifier();
+        return PlanetResource::collection(DB::table('planets')
+            ->join('solar_systems', 'planets.solar_system_id', '=', 'solar_systems.id')
+            ->join('galaxies', 'solar_systems.galaxy_id', '=','galaxies.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('planets.deleted_at', '=', null)
+            ->select('planets.*')
+            ->get());
     }
 
     /**
@@ -28,9 +35,24 @@ class PlanetController extends Controller
      */
     public function store(StorePlanetRequest $request, $solarSystemId)
     {
-        $planet = Planet::create($request->all());
-
-        return response()->json(['status' => 'Planeta criado com sucesso!', new PlanetResource($planet)]);
+        $userId = auth('api')->user()->getAuthIdentifier();
+        $solarSystemBelongsToUser = DB::table('users')
+            ->join('galaxies', 'galaxies.user_id', '=', 'users.id')
+            ->join('solar_systems', 'solar_systems.galaxy_id', '=','galaxies.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('solar_systems.id', '=', $solarSystemId)
+            ->where('solar_systems.deleted_at', '=', null)
+            ->select('solar_systems.*')
+            ->first();
+        if($solarSystemBelongsToUser != null) {
+            $data = ($request->validated());
+            $data['solar_system_id'] = $solarSystemId;
+            $planet = Planet::create($data);
+            return response()->json(['status' => 'Planeta criado com sucesso!', new PlanetResource($planet)]);
+        }
+        else {
+            return response()->json(['status' => 'Não é possível criar esse Planeta!']);
+        }
     }
 
     /**
@@ -41,7 +63,22 @@ class PlanetController extends Controller
      */
     public function show($id)
     {
-        return new PlanetResource(Planet::FindOrFail($id));
+        $userId = auth('api')->user()->getAuthIdentifier();
+        $planetBelongsToUser = DB::table('users')
+            ->join('galaxies', 'galaxies.user_id', '=', 'users.id')
+            ->join('solar_systems', 'solar_systems.galaxy_id', '=','galaxies.id')
+            ->join('planets', 'planets.solar_system_id', '=', 'solar_systems.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('planets.id', '=', $id)
+            ->where('planets.deleted_at', '=', null)
+            ->select('planets.*')
+            ->first();
+        if($planetBelongsToUser != null) {
+            return new PlanetResource(Planet::FindOrFail($id));
+        }
+        else {
+            return response()->json(['status' => 'Busca não retornou resultados.']);
+        }
     }
 
     /**
@@ -53,11 +90,24 @@ class PlanetController extends Controller
      */
     public function update(StorePlanetRequest $request, $id)
     {
-        $planet = Planet::findOrFail($id);
-
-        $planet->update($request->all());
-
-        return response()->json(['status' => 'Informações sobre Planeta atualizadas com sucesso!', new PlanetResource($planet)]);
+        $userId = auth('api')->user()->getAuthIdentifier();
+        $planetBelongsToUser = DB::table('users')
+            ->join('galaxies', 'galaxies.user_id', '=', 'users.id')
+            ->join('solar_systems', 'solar_systems.galaxy_id', '=','galaxies.id')
+            ->join('planets', 'planets.solar_system_id', '=', 'solar_systems.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('planets.id', '=', $id)
+            ->where('planets.deleted_at', '=', null)
+            ->select('planets.*')
+            ->first();
+        if($planetBelongsToUser != null) {
+            $planet = Planet::findOrFail($id);
+            $planet->update($request->all());
+            return response()->json(['status' => 'Informações sobre Planeta atualizadas com sucesso!', new PlanetResource($planet)]);
+        }
+        else {
+            return response()->json(['status' => 'Não é possível autualizar as informações desse Planeta!']);
+        }
     }
 
     /**
@@ -68,8 +118,23 @@ class PlanetController extends Controller
      */
     public function destroy($id)
     {
-        $planet = Planet::findOrFail($id);
-        $planet->delete();
-        return response()->json(['status' => 'Planeta excluído com sucesso!']);
+        $userId = auth('api')->user()->getAuthIdentifier();
+        $planetBelongsToUser = DB::table('users')
+            ->join('galaxies', 'galaxies.user_id', '=', 'users.id')
+            ->join('solar_systems', 'solar_systems.galaxy_id', '=','galaxies.id')
+            ->join('planets', 'planets.solar_system_id', '=', 'solar_systems.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('planets.id', '=', $id)
+            ->where('planets.deleted_at', '=', null)
+            ->select('planets.*')
+            ->first();
+        if($planetBelongsToUser != null) {
+            $planet = Planet::findOrFail($id);
+            $planet->delete();
+            return response()->json(['status' => 'Planeta excluído com sucesso!']);
+        }
+        else {
+            return response()->json(['status' => 'Não é possível excluir esse Planeta!']);
+        }
     }
 }
