+++ galaxies/app/Http/Controllers/Api/SolarSystemController.php
@@ -5,9 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreSolarSystemRequest;
 use App\Http\Resources\SolarSystemResource;
-use App\Models\Planet;
 use App\Models\SolarSystem;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
 class SolarSystemController extends Controller
@@ -19,7 +17,13 @@ class SolarSystemController extends Controller
      */
     public function index()
     {
-        return SolarSystemResource::collection(SolarSystem::all());
+        $userId = auth('api')->user()->getAuthIdentifier();
+        return SolarSystemResource::collection(DB::table('solar_systems')
+            ->join('galaxies', 'solar_systems.galaxy_id', '=','galaxies.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('solar_systems.deleted_at', '=', null)
+            ->select('solar_systems.*')
+            ->get());
     }
 
     /**
@@ -30,8 +34,23 @@ class SolarSystemController extends Controller
      */
     public function store(StoreSolarSystemRequest $request, $galaxyId)
     {
-        $solarSystem = SolarSystem::create($request->all());
-        return response()->json(['status' => 'Sistema Solar criado com sucesso!', new SolarSystemResource($solarSystem)]);
+        $userId = auth('api')->user()->getAuthIdentifier();
+        $galaxyBelongsToUser = DB::table('users')
+            ->join('galaxies', 'galaxies.user_id', '=', 'users.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('galaxies.id', '=', $galaxyId)
+            ->where('galaxies.deleted_at', '=', null)
+            ->select('galaxies.*')
+            ->first();
+        if($galaxyBelongsToUser != null) {
+            $data = ($request->validated());
+            $data['galaxy_id'] = $galaxyId;
+            $solarSystem = SolarSystem::create($data);
+            return response()->json(['status' => 'Sistema Solar criado com sucesso!', new SolarSystemResource($solarSystem)]);
+        }
+        else {
+            return response()->json(['status' => 'Não é possível criar esse Sistema Solar!']);
+        }
     }
 
     /**
@@ -42,7 +61,21 @@ class SolarSystemController extends Controller
      */
     public function show($id)
     {
-        return new SolarSystemResource(SolarSystem::FindOrFail($id));
+        $userId = auth('api')->user()->getAuthIdentifier();
+        $solarSystemBelongsToUser = DB::table('users')
+            ->join('galaxies', 'galaxies.user_id', '=', 'users.id')
+            ->join('solar_systems', 'solar_systems.galaxy_id', '=','galaxies.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('solar_systems.id', '=', $id)
+            ->where('solar_systems.deleted_at', '=', null)
+            ->select('solar_systems.*')
+            ->first();
+        if($solarSystemBelongsToUser != null) {
+            return new SolarSystemResource(SolarSystem::FindOrFail($id));
+        }
+        else {
+            return response()->json(['status' => 'Busca não retornou resultados.']);
+        }
     }
 
     /**
@@ -54,10 +87,23 @@ class SolarSystemController extends Controller
      */
     public function update(StoreSolarSystemRequest $request, $id)
     {
-        $solarSystem = SolarSystem::findOrFail($id);
-
-        $solarSystem->update($request->all());
-        return response()->json(['status' => 'Informações sobre Sistema Solar atualizadas com sucesso!', new SolarSystemResource($solarSystem)]);
+        $userId = auth('api')->user()->getAuthIdentifier();
+        $solarSystemBelongsToUser = DB::table('users')
+            ->join('galaxies', 'galaxies.user_id', '=', 'users.id')
+            ->join('solar_systems', 'solar_systems.galaxy_id', '=','galaxies.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('solar_systems.id', '=', $id)
+            ->where('solar_systems.deleted_at', '=', null)
+            ->select('solar_systems.*')
+            ->first();
+        if($solarSystemBelongsToUser != null) {
+            $solarSystem = SolarSystem::findOrFail($id);
+            $solarSystem->update($request->all());
+            return response()->json(['status' => 'Informações sobre Sistema Solar atualizadas com sucesso!', new SolarSystemResource($solarSystem)]);
+        }
+        else {
+            return response()->json(['status' => 'Não é possível autualizar as informações desse Sistema Solar!']);
+        }
     }
 
     /**
@@ -68,16 +114,22 @@ class SolarSystemController extends Controller
      */
     public function destroy($id)
     {
-        $solarSystem = SolarSystem::findOrFail($id);
-        /*$planets = DB::table('planets')
-            ->where('solar_system_id', '=', $id)->get();*/
-
-        $solarSystem->delete();
-
-       /* if(isset($planets)) {
-            $planets->delete();
-        }*/
-
-        return response()->json(['status' => 'Sistema Solar excluído com sucesso!']);
+        $userId = auth('api')->user()->getAuthIdentifier();
+        $solarSystemBelongsToUser = DB::table('users')
+            ->join('galaxies', 'galaxies.user_id', '=', 'users.id')
+            ->join('solar_systems', 'solar_systems.galaxy_id', '=','galaxies.id')
+            ->where('galaxies.user_id', '=', $userId)
+            ->where('solar_systems.id', '=', $id)
+            ->where('solar_systems.deleted_at', '=', null)
+            ->select('solar_systems.*')
+            ->first();
+        if($solarSystemBelongsToUser != null) {
+            $solarSystem = SolarSystem::findOrFail($id);
+            $solarSystem->delete();
+            return response()->json(['status' => 'Sistema Solar excluído com sucesso!']);
+        }
+        else {
+            return response()->json(['status' => 'Não é possível excluir esse Sistema Solar!']);
+        }
     }
 }
